+++ xradar/io/backends/__init__.py
@@ -10,11 +10,17 @@ Xarray Backends
     :maxdepth: 4
 
 .. automodule:: xradar.io.backends.cfradial1
+.. automodule:: xradar.io.backends.gamic
 .. automodule:: xradar.io.backends.odim
+.. automodule:: xradar.io.backends.furuno
+.. automodule:: xradar.io.backends.rainbow
 
 """
 
 from .cfradial1 import *  # noqa
+from .furuno import *  # noqa
+from .gamic import *  # noqa
 from .odim import *  # noqa
+from .rainbow import *  # noqa
 
 __all__ = [s for s in dir() if not s.startswith("_")]